chore: group the code in util.py
