+++ neosca/util_platform_info.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# -*- coding=utf-8 -*-
+
+import sys
+import os
+IS_WINDOWS = sys.platform == "win32"
+IS_DARWIN = sys.platform == "darwin"
+IS_LINUX = sys.platform == "linux"
